% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveDriver.R
\docType{data}
\name{kinase_domains}
\alias{kinase_domains}
\title{Example kinase domains for ActiveDriver}
\format{A data frame with 1 observation of 4 variables}
\usage{
data(ActiveDriver_data)
}
\description{
A dataset describing kinase domains. The variables are as follows:
}
\details{
\itemize{
  \item gene. the gene symbol of the gene where the kinase domain occurs
  \item position. the position in the protein sequence where the kinase domain begins
  \item phos. TRUE
  \item residue. the kinase domain residues
}
}
\keyword{datasets}
