% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveDriver.R
\docType{data}
\name{mutations}
\alias{mutations}
\title{Example mutations for ActiveDriver}
\format{A data frame with 408 observations of 5 variables}
\usage{
data(ActiveDriver_data)
}
\description{
A dataset describing mis-sense mutations (i.e., substitutions in proteins). The variables are as follows:
}
\details{
\itemize{
  \item gene. the mutated gene
  \item sample_id. the sample where the mutation originates
  \item position. the position in the protein sequence where the mutation occurs
  \item wt_residue. the wild-type residue
  \item mut_residue. the mutant residue
}
}
\keyword{datasets}
