% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveDriver.R
\docType{data}
\name{phosphosites}
\alias{phosphosites}
\title{Example phosphosites for ActiveDriver}
\format{A data frame with 131 observations of 4 variables}
\usage{
data(ActiveDriver_data)
}
\description{
A dataset describing protein phosphorylation sites. The variables are as follows:
}
\details{
\itemize{
  \item gene. the gene symbol the phosphosite occurs in
  \item position. the position in the protein sequence where the phosphosite occurs
  \item residue. the phosphosite residue
  \item kinase. the kinase that phosphorylates this site
}
}
\keyword{datasets}
