% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveDriver.R
\name{read_fasta}
\alias{read_fasta}
\title{Read FASTA file as character vector.}
\usage{
read_fasta(fname)
}
\arguments{
\item{fname}{name of file to be read.}
}
\value{
character vector with names corresponding to annotations from FASTA.
}
\description{
Read FASTA file as character vector.
}
