% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveDriver.R
\docType{data}
\name{sequence_disorder}
\alias{sequence_disorder}
\title{Example protein disorder for ActiveDriver}
\format{A named character vector with 4 elements}
\usage{
data(ActiveDriver_data)
}
\description{
A dataset containing the disorder of four proteins.
}
\keyword{datasets}
