% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveDriver.R
\docType{data}
\name{sequences}
\alias{sequences}
\title{Example protein sequences for ActiveDriver}
\format{A named character vector with 4 elements}
\usage{
data(ActiveDriver_data)
}
\description{
A dataset containing the sequences of four proteins.
}
\keyword{datasets}
