\name{read_fasta}
\alias{read_fasta}
\title{Read FASTA file as character vector.}
\usage{
  read_fasta(fname)
}
\arguments{
  \item{fname}{name of file to be read.}
}
\value{
  character vector with names corresponding to annotations
  from FASTA.
}
\description{
  Read FASTA file as character vector.
}

